+++ R/zzz.R
@@ -2,7 +2,16 @@
 #' @importFrom knitr kable
 #' @importFrom glue glue
 #' @importFrom yaml yaml.load
-#' @importFrom utils head tail 
+#' @importFrom utils head tail
 #'
 #' @keywords internal
 NULL
+
+
+
+
+# catAnim1 <- function(k, ...) {
+#   symb <- cli::symbol[c(45:53, 52:46)]
+#   nb <- length(symb)
+#   cat(symb[[1 + (k %% nb)]], ..., "\r")
+# }