+++ tests/testthat/test-stopwatch.R
@@ -1,0 +1,6 @@
+context("Stopwatch & Timer")
+
+test_that("messages", {
+  expect_message(stopwatch(.1))
+  expect_message(timer(.1))
+})