+++ NEWS.md
@@ -1,5 +1,6 @@
-# inSilecoMisc 0.4.0 (devel)
+# inSilecoMisc (devel)
 
+* add `stopwatch()` and `timer()`.
 * add `seqRg()` that generates a regular sequence based on the range of a vector.
 * add `seqRow()` and ``seqCol()` that generate a regular sequences based on the number of rows and columns (respectively) of a data frame (or a matrix).